Move ntk computation to analysis
