+++ CI/unit_tests/analysis/test_jax_ntk.py
@@ -31,8 +31,8 @@ import optax
 from flax import linen as nn
 from jax import random
 
+from znnl.analysis import JAXNTKComputation
 from znnl.models import FlaxModel
-from znnl.ntk_computation import JAXNTKComputation
 
 
 class FlaxTestModule(nn.Module):
