+++ CI/unit_tests/analysis/test_jax_ntk_subsampling.py
@@ -30,8 +30,8 @@ import optax
 from flax import linen as nn
 from jax import random
 
+from znnl.analysis import JAXNTKSubsampling
 from znnl.models import FlaxModel
-from znnl.ntk_computation import JAXNTKSubsampling
 
 
 class FlaxTestModule(nn.Module):
