+++ CI/unit_tests/models/_test_huggingface_flax_model.py
@@ -29,8 +29,8 @@ import optax
 from jax import random
 from transformers import FlaxResNetForImageClassification, ResNetConfig
 
+from znnl.analysis import JAXNTKComputation
 from znnl.models import HuggingFaceFlaxModel
-from znnl.ntk_computation import JAXNTKComputation
 
 
 class TestFlaxHFModule:
