+++ CI/unit_tests/models/test_flax_model.py
@@ -34,8 +34,8 @@ import optax
 from flax import linen as nn
 from jax import random
 
+from znnl.analysis import JAXNTKComputation
 from znnl.models import FlaxModel
-from znnl.ntk_computation import JAXNTKComputation
 
 
 class FlaxTestModule(nn.Module):
